% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predInterval.R
\name{predInterval}
\alias{predInterval}
\title{Conformal prediction inference for survival data with truncation}
\usage{
predInterval(
  XTest,
  ATr,
  YTr,
  XTr,
  del = NULL,
  cens = TRUE,
  alp = 0.1,
  model = "cox",
  tau = NULL
)
}
\arguments{
\item{XTest}{a matrix of covariates with each row corresponding to an instance of covariate vector.}

\item{ATr}{truncation time.}

\item{YTr}{the observed event time.}

\item{XTr}{the observed covariate matrix}

\item{del}{the censoring indicator vector.}

\item{cens}{the censoring existence indicator.}

\item{alp}{miscoverage probability.}

\item{model}{a model name to indicate which model will be used to fit the data.}

\item{tau}{the restricted maximum time.}
}
\description{
The predInterval function produces conformal prediction intervals for survival time given the coverates.
}
