% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightFunction.R
\name{survTrunPL}
\alias{survTrunPL}
\title{Generate the product limiting estimator for survival functions, given a, y, and label, d.}
\usage{
survTrunPL(a, y, d)
}
\description{
Generate the product limiting estimator for survival functions, given a, y, and label, d.
}
