% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightFunction.R
\name{survTrunPLNoCens}
\alias{survTrunPLNoCens}
\title{Generate the product limiting estimator for survival functions, given a, y, and label, d.}
\usage{
survTrunPLNoCens(a, y)
}
\description{
Generate the product limiting estimator for survival functions, given a, y, and label, d.
}
