% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightFunction.R
\name{weightFunFact}
\alias{weightFunFact}
\title{Generate the product limiting estimator for survival functions, given a, y, and label, d.}
\usage{
weightFunFact(y, a, d = NULL)
}
\description{
Generate the product limiting estimator for survival functions, given a, y, and label, d.
}
